backend/agents/analyze-agent/requirement.txt
flask==2.3.3
google-cloud-firestore==2.11.1
google-cloud-pubsub==2.18.1
google-cloud-translate==3.11.3
google-genai
googlemaps==4.10.0
firebase-admin==6.2.0
tweepy==4.14.0
feedparser==6.0.10
requests==2.31.0
python-dateutil==2.8.2
pymongo
requests
motor